Copy URL and tooltips for URL card btns
