+++ src/static/scripts/components/urls_deck/access_all_urls_btn.js
@@ -65,6 +65,6 @@ function accessAllURLsInUTub() {
   const visibleURLsToAccess = $.map(visibleURLs, (url) => $(url).attr("href"));
 
   for (let i = 0; i < visibleURLsToAccess.length; i++) {
-    accessLink(visibleURLsToAccess[i]);
+    accessLink(visibleURLsToAccess[i], null);
   }
 }
