+++ src/static/scripts/components/urls_deck/url_cards/url_cards_options_btns/url_cards_btns.js
@@ -1,0 +1,26 @@
+"use strict";
+
+// Create all the buttons necessary for a url card
+function createURLOptionsButtons(url, urlCard) {
+  const urlOptions = $(document.createElement("div")).addClass(
+    "urlOptions justify-content-start flex-row gap-15p",
+  );
+
+  urlOptions
+    .append(createAccessLinkBtn(url))
+    .append(createAddTagBtn(urlCard))
+    .append(createCopyURLBtn(url));
+
+  if (url.canDelete) {
+    urlOptions
+      .append(createEditURLBtn(urlCard))
+      .append(createDeleteURLBtn(url, urlCard));
+  }
+
+  const urlCardLoadingIcon = $(document.createElement("div")).addClass(
+    "urlCardDualLoadingRing",
+  );
+  urlOptions.append(urlCardLoadingIcon);
+
+  return urlOptions;
+}
