+++ src/templates/_strings.html
@@ -16,6 +16,13 @@ class Strings {
 			this.UTUB_SELECT = "{{CONSTANTS.STRINGS.UTUB_SELECT}}"
 			this.FORGOT_YOUR_PASSWORD = "{{CONSTANTS.STRINGS.FORGOT_YOUR_PASSWORD}}"
 			this.LOGIN_TITLE = "{{CONSTANTS.STRINGS.LOGIN_TITLE}}"
+			this.COPY_URL_TOOLTIP = "{{CONSTANTS.STRINGS.COPY_URL_TOOLTIP}}"
+			this.COPIED_URL_TOOLTIP = "{{CONSTANTS.STRINGS.COPIED_URL_TOOLTIP}}"
+			this.COPIED_URL_FAILURE_TOOLIP = "{{CONSTANTS.STRINGS.COPIED_URL_FAILURE_TOOLIP}}"
+			this.ACCESS_URL_TOOLTIP = "{{CONSTANTS.STRINGS.ACCESS_URL_TOOLTIP}}"
+			this.ADD_URL_TAG_TOOLTIP = "{{CONSTANTS.STRINGS.ADD_URL_TAG_TOOLTIP}}"
+			this.EDIT_URL_TOOLTIP = "{{CONSTANTS.STRINGS.EDIT_URL_TOOLTIP}}"
+			this.DELETE_URL_TOOLTIP = "{{CONSTANTS.STRINGS.DELETE_URL_TOOLTIP}}"
 			Strings.instance = this;
 		}
