+++ src/templates/home/UTubDeck/UTubDeckHeaders.html
@@ -57,7 +57,7 @@
         <div id="SearchUTubWrap" class="input-text-holder flex-column createDiv hidden">
             <div class="text-input-container">
                 <div class="text-input-inner-container">
-                    <input placeholder="Search UTubs" class="text-input focus search-input" type="text" id="UTubNameSearch" name="utubName" minlength="{{CONSTANTS.UTUBS.MIN_NAME_LENGTH}}" maxlength="{{CONSTANTS.UTUBS.MAX_NAME_LENGTH}}" required="" tabindex="0">
+                    <input placeholder="Search UTubs" class="text-input search-input" type="text" id="UTubNameSearch" name="utubName" minlength="{{CONSTANTS.UTUBS.MIN_NAME_LENGTH}}" maxlength="{{CONSTANTS.UTUBS.MAX_NAME_LENGTH}}" required="" tabindex="0">
                     <label class="text-input-label search-input-label top-50perc left-10p font-size-16p" for="utubNameSearch">Search UTubs</label>
                 </div>
             </div>
