+++ src/utils/constants.py
@@ -4,7 +4,16 @@ from src.utils.strings.email_validation_strs import (
     VALIDATE_YOUR_EMAIL,
 )
 from src.utils.strings.splash_form_strs import FORGOT_YOUR_PASSWORD, LOGIN_TITLE
-from src.utils.strings.url_strs import DELETE_URL_WARNING
+from src.utils.strings.tag_strs import ADD_URL_TAG_TOOLTIP
+from src.utils.strings.url_strs import (
+    ACCESS_URL_TOOLTIP,
+    COPIED_URL_FAILURE_TOOLIP,
+    COPIED_URL_TOOLTIP,
+    COPY_URL_TOOLTIP,
+    DELETE_URL_TOOLTIP,
+    DELETE_URL_WARNING,
+    EDIT_URL_TOOLTIP,
+)
 from src.utils.strings.user_strs import MEMBER_DELETE_WARNING, MEMBER_LEAVE_WARNING
 from src.utils.strings.utub_strs import (
     UTUB_CREATE_MSG,
@@ -73,6 +82,13 @@ class STRINGS:
     UTUB_SELECT = UTUB_SELECT
     FORGOT_YOUR_PASSWORD = FORGOT_YOUR_PASSWORD
     LOGIN_TITLE = LOGIN_TITLE
+    COPY_URL_TOOLTIP = COPY_URL_TOOLTIP
+    COPIED_URL_TOOLTIP = COPIED_URL_TOOLTIP
+    COPIED_URL_FAILURE_TOOLIP = COPIED_URL_FAILURE_TOOLIP
+    ACCESS_URL_TOOLTIP = ACCESS_URL_TOOLTIP
+    ADD_URL_TAG_TOOLTIP = ADD_URL_TAG_TOOLTIP
+    EDIT_URL_TOOLTIP = EDIT_URL_TOOLTIP
+    DELETE_URL_TOOLTIP = DELETE_URL_TOOLTIP
 
 
 class CONSTANTS:
