+++ src/utils/strings/url_strs.py
@@ -23,6 +23,12 @@ URL_MODIFIED = "URL modified."
 URL_FOUND_IN_UTUB = "URL found in UTub."
 TAG_IN_UTUB = "tagInUTub"
 DELETE_URL_WARNING = "You can always add it back again!"
+COPY_URL_TOOLTIP = "Copy URL"
+COPIED_URL_TOOLTIP = "Copied!"
+COPIED_URL_FAILURE_TOOLIP = "Copy failed!"
+ACCESS_URL_TOOLTIP = "Access URL"
+EDIT_URL_TOOLTIP = "Edit URL"
+DELETE_URL_TOOLTIP = "Delete URL"
 
 
 class URL_GENERAL:
